Fix stack trace if vendordata_raw contained an array (#837)
The implementation in existing datasources means that vendordata_raw is
not "raw" as it ideally would be. Instead, actual values may include
bytes, string or list.  If the value was a list, then the attempt to
persist that data to a file in '_store_rawdata' would raise a
TypeError.

The change is to encode with util.json_dumps (which is safe for
binary data) before writing.