patcher.py: drop imp
The imp module is deprecated and replaced by importlib, which according
to the documentation has no replacement for acquire_lock() and
release_lock(), which are the only reason why `imp` is imported.

I see nothing about the code using this lock that actually requires it.
Let's remove the locking code and the import altogether.