Add module to allow deferring file writing
THe main idea is to introduce a second module that takes care of
writing files, but in the 'final' stage.

While the introduction of a second module would allow for choosing
the appropriate place withing the order of modules (and stages),
there is no addition top-level directive being added to the cloud
configuration schema. Instead, existing schemes are being extended,
like the 'users' directive.

The new module 'write-deferred-files' tries to reuse as much as
possible of the 'write-files' functionality, including the scheme
or the 'write_files' function.

further details on how it would be used, are shown in the scheme
definition of the module.