Consider folder permission when chosing user_authorizedkeys_file in extract_authorized_keys
In /etc/ssh/sshd_config, it is possible to define a custom authorized_keys file
that will contain the keys allowed to access the machine via
the AuthorizedKeysFile option.
Cloudinit is able to add user-specific keys to the existing ones, but we need
to be careful on which of the authorized_keys files listed to pick.
Chosing for example a file that is shared by all user will cause security
issues, because the owner of that key can then access also other users.

We therefore pick an authorized_keys file only if it satisfies the following
conditions:
1. it is not a "global" file, ie it must be defined in AuthorizedKeysFile
	with %u, %h or be in  /home/<user>. This avoids security issues.
2. it must comply with ssh permission requirements, otherwise the ssh agent
	won't use that file.

We also need to consider the case when the chosen authorized_keys file
does not exist. In this case, the existing behavior of cloud-init is
to create the new file. We therefore need to be sure that the file complies
with ssh permissions too, by setting:
- the actual file to permission 600, and owned by the user
- the directories in the path that do not exist must be root owned and with
	permission 755.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>