cc_update_etc_hosts: Use the distribution-defined path for the hosts file
The distribution class has a field that specifies the location of
the system hosts file and this can be overridden in subclasses.
While the field is correctly used in distro.update_etc_hosts(), the
update_etc_hosts module does not use it and just assumes '/etc/hosts'

This fixes the module to use the distribution-specific variable.