ssh_utils.py: ignore when sshd_config options are not key/value pairs
As specified in #LP 1845552,
In cloudinit/ssh_util.py, in parse_ssh_config_lines(), we attempt to
parse each line of sshd_config. This function expects each line to
be one of the following forms:

    \# comment
    key value
    key=value

However, options like DenyGroups and DenyUsers are specified to
*optionally* accepts values in sshd_config.
Cloud-init should comply to this and skip the option if a value
is not provided.

Signed-off-by: Emanuele Giuseppe Esposito <eesposit@redhat.com>