cc_ssh_authkey_fingerprints.py: prevent duplicate messages on console
When cloud-init is configured to show SSH user key fingerprints during
boot two of the same message appears for each user. This appears to be as
the util.multi_log call defaults to send to both console directly and to
stderr (which also goes to console).

This change sends them only to console directly.