Some unittest isn't cleaning up resources correctly - work around that
Patching subp is failing for these tests - successful when pytest is
executed individually against this test, but when run against all the tests
it fails.

This test validates the same logic but mocks the function that calls
subp instead. It might be test_util.py that is not cleaning up
correctly, but I can't say for certain.

(test_util.py might be the culprit?)