lxd: provide v1.0 prefix for API version to aid in jinja alias
Without a leading non-integer character jinja's lexer gets confused trying to
parse underscore-delimited variable names such as 1_0 as the lexer starts to
interpret an integer and chokes on the underscore. This prevents cloud-init
from providing a simple jinja dot-notation variable alias such as ds.1_0
because the lexer barfs claiminig invalid variable end "_0".

Providing the leading character allows the lexer to interpret the variable
as a string type and properly interpret v1_0 as the whole variable text.