sources/azure: do not persist failed_desired_api_version flag
If get_imds_data_with_api_fallback() falls back to the minimum required
API version, it is effectively pinned to the old API version forever.

Remove the failed_desired_api_version property to prevent persistence of
the flag between calls and/or reboots.

The continued presence of this flag in obj.pkl should be harmless.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>