sources/azure: remove lease file parsing for wireserver IP
Always review the ephemeral DHCP lease when performing to find
the wireserver endpoint specified with option 245.

DHCP option 245 is guaranteed to be provided, which we will now
immediately parse when bringing up ephemeral networking with
_process_dhcp_lease(),  saving the parsed address as
_wireserver_addres (which safely defaults to the expected IP
address if the option is not found for whatever reason).

In practice, the fallback parsing of lease files has never been
necessary because the wireserver IP has never changed to date.

This allows us to safely remove the dhclient_lease_file option
and reduce some complexity through consolidation of lease
handling.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>