net: introduce find_candidate_nics()
find_fallback_nic_on_linux(), etc. provides valuable filtering of
network interfaces in an effort to determine the best candidate
for the fallback interface.

Expose this logic with a new set of methods for finding the candidate
network interfaces.  These methods can be used by data sources which
cannot rely on the fallback interface being the correct choice.

Note that the MAC address filtering is now part of
find_candidate_nics_on_linux().  This should be consistent behavior
as find_fallback_nic_on_linux() never selected an interface without
a MAC.  find_fallback_nic_on_linux() continues to prefer eth0,
but we make no such distinction in the candidate search.

Specifically, for Linux, order the NICs:

1. DEFAULT_PRIMARY_INTERFACE, if connected.
2. Remaining connected interfaces, naturally sorted.
3. DEFAULT_PRIMARY_INTERFACE, if possibly connected.
4. Remaining possibly connected interfaces, naturally sorted.

This way the fallback calculation is consistent with previous
behavior, but we don't ignore possibly connected interfaces just
because others may be connected.

For non-Linux, continue to sort interfaces naturally.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>