Revert "net: Make sysconfig renderer compatible with Network Manager."
Firstly, this relies upon the fact that you can get ifcfg support by adding
it to NetworkManager.conf. That is not guarranteed and certianly will not
be case in future.

Secondly, cloud-init always generates configuration with
NM_CONTROLLED=no, so the generated ifcfg files are no good for
NetworkManager. Fedora patches around this by just removing those lines
in their cloud-init package.

Let's remove this and add a proper NetworkManager support later on
instead.

This reverts commit 3861102fcaf47a882516d8b6daab518308eb3086.