testing: stop universally overwriting /etc/cloud/cloud.cfg.d
As part of IN_PLACE testing, /etc/cloud/cloud.cfg.d get overwritten by
what's in the source tree. This can cause problems when the directory is
mounted in, because tests need the ability to modify files in
/etc/cloud. Attempting to 'lxc file push'  instead will fail on LXD VMs
because the LXD agent isn't available yet.

If such functionality is desired, one can temporarily use the
'lxd_setup' mark while writing a test and push the files manually.