sources/azure: removed unused savable PPS paths
If the VM is rebooted during provisioning, the PPS type will be
determined to be UNKNOWN and will poll for reprovision data.

Given that we will never enter _wait_for_all_nics_ready() in any
other condition than a fresh source instance in Savable PPS, we can
safely remove the now-unused code paths.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>