Better interface handling for Vultr, expect unexpected DHCP servers (#1297)
If a private interface is configured, has a DHCP server on it, and
cloud-init runs, the datasource fails to fail over to the next
interface interrupting seeking. Configuring it as such can also lead
to delays in deploy time if it did. This sorts the interface list to
limit the potential of this occurring. Checks for the metadata route
before trying, and now appropriately fails over if those contingencies
fails by appropriately continuing to seek on metadata failure as was
originally intended. Ideally the upcoming candidate_nic would have
been used here to avoid the interface list function.