Add --no-update to add-apt-repostory call
In cc_apt_configure, "add-apt-repository" is called a few lines before
we make a call to update packages, so the additional update that
happens in "add-apt-repository" is redundant and a waste of time.
Additionally, the update in add-apt-repository can fail on
obtaining apt lock and cause the module to raise an exception.