Reworking dual_stack for more robust connection handling:
This allows a ConnectionError exception to be thrown in one of the
threads without canceling the other thread - the first successful call
is returned, the rest are canceled. If all threads have exceptions,
log them all and raise the last one (need choose one somehow, right? I
chose the last; tell me if there is a reason we should raise the first
in order of execution or first in alphabetical order or whatever)

Tests to verify the above described behavior have been added.