tox: integration-tests-jenkins: softfail if only some test failed
Prior to this change a single test failure when running the
integration-tests-jenkins tox environment caused Jenkins jobs to fully
fail (status FAILURE).

With this change the pytest's "some tests failed" return code [1] is
ignored, and the status of the Jenkins jobs is determined using the
junit report that pytest generates. Most of the time this will result
in UNSTABLE (yellow) jobs when some tests failed.

If the junit report is missing (or only reports failures) the job
will still be marked as FAILURE (red).

[1] https://docs.pytest.org/en/7.1.x/reference/exit-codes.html