Improve DataSourceOVF error reporting when script disabled (#1525)
In GUESTCUST_ERROR_SCRIPT_DISABLED failure, cloud-init reports
customization state (and error code) twice, which are used to be
(4, 6) and (4, 100). The interval of the two reporting is short.
Sometimes the first reporting (4, 6) can be inhibited, which may make
users confused about the customization failure reason.
This update changes the second reporting from (4, 100) to (4, 6).
This change is compatible with the VMware traditional guest OS
customization behavior.