tests: add test for parsing static dns for existing devices (#1557)
This patch adds a test case for the launchpad bug below. When parsing a
network configuration where a device on the system has the same mac
address as given in the `match` section of the config, but with the real
device name not matching the config name, parsing of a static nameserver
entry fails with a `KeyError`.

Validates behavior in af404781299 which fixes LP #197877.